SnapdragonLauncher: add and remove screen for FlatLauncher
Add function for Flat-Launcher: add new screen and remove empty screen.
Description: add a new page which only contains a button after last one in overview mode,
it would add a new empty page at last one when the button was clicked,and each empty
page could be deleted as long as the it's delete button was clicked.

Change-Id: Ie23ac9565dd86018c2b700274bdc23e73b1c4037