SnapdragonLauncher: Fix view requestFocus NullpointerException
When remove all apps from final page to other pages or
uninstall them,the final page will be deleted,but when
only one page,uninstall all apps in this page would also
delete the page,so workspace children count will be equal
to 0 and lead to NullpointerException.

The method of delete page shouldn't be executed when only
one page.

Change-Id: I06184825badc015d662a58ac4481e84b5284d4c3
CRs-Fixed: 1060430