SnapdragonLauncher: Display two shortcut APP icons
Go to Settings and set any one of app which in Dock
to Disabled,the app icon doesn't disappear.The method
of remove items only works in Workspace but Dock because
of the Dock structure redesign.

Traversal of all items in Dock so remove method will work
on all of items on desktop.

Change-Id: Iac08faf694e0f9e9376a914e2646e09321f8786b
CRs-Fixed: 1063301