SnapdragonLauncher:  Crash when put contact widget to folder
Description:
Contact widget is a special launcher icon. It is only
shortcut icon and it doesn't have the component. Folder
need to get the unread Number of app via component, so
it leads to NullPointer exception if component is null.

Check the component of shortcut icon whether is null.

Change-Id: Id692383ba8a6ab22f39ebec119b942677a54d003
CRs-Fixed: 1076348