SnapdragonLauncher: Fix monkey crash
Attempt to invoke virtual method 'void android.view.View.
addFocusables(java.util.ArrayList, int, int)' on a null
object reference.

When workspace status changed and adding page is null,
it's not reasonable to update the order of the workspace
screens in the database.
To ensure that current page is valid page.

Change-Id: I52e8c7cc88ffb68a73632c91fb47f621736d7ff8
CRs-Fixed: 1077421