SnapdragonLauncher: Fix display empty space in hotseat
The visibility of child view would be set to invisible
when item drag from external, at this time, drag view
to the other child to create folder or add to existing
folder, such an operation doesn't execute onDrop method
so that the empty view before will not be set to gone.

Set the visibility of before invisible view to gone
after above operation.

Change-Id: I91c6482f44dec9ed01d37adaa6abdc37689a33a6
CRs-Fixed: 1087063